requirements.txt
python-dotenv~=1.1.1
langchain~=0.3.27
langchain-community~=0.3.29
elevenlabs
playsound
pygame
mistralai
langchain-mistralai
sentence-transformers